decent connection status
