+++ ui/litellm-dashboard/src/components/mcp_tools/create_mcp_server.tsx
@@ -12,6 +12,7 @@ import { Button, TextInput } from "@tremor/react";
 import { createMCPServer } from "../networking";
 import { MCPServer, MCPServerCostInfo } from "./types";
 import MCPServerCostConfig from "./mcp_server_cost_config";
+import MCPConnectionStatus from "./mcp_connection_status";
 import { isAdminRole } from "@/utils/roles";
 
 
@@ -33,6 +34,7 @@ const CreateMCPServer: React.FC<CreateMCPServerProps> = ({
   const [isLoading, setIsLoading] = useState(false);
   const [costConfig, setCostConfig] = useState<MCPServerCostInfo>({});
   const [formValues, setFormValues] = useState<Record<string, any>>({});
+  const [tools, setTools] = useState<any[]>([]);
 
   const handleCreate = async (formValues: Record<string, any>) => {
     setIsLoading(true);
@@ -58,6 +60,7 @@ const CreateMCPServer: React.FC<CreateMCPServerProps> = ({
         message.success("MCP Server created successfully");
         form.resetFields();
         setCostConfig({});
+        setTools([]);
         setModalVisible(false);
         onCreateSuccess(response);
       }
@@ -74,6 +77,7 @@ const CreateMCPServer: React.FC<CreateMCPServerProps> = ({
   const handleCancel = () => {
     form.resetFields();
     setCostConfig({});
+    setTools([]);
     setModalVisible(false);
   };
 
@@ -250,14 +254,22 @@ const CreateMCPServer: React.FC<CreateMCPServerProps> = ({
               </Form.Item>
             </div>
 
-            {/* Cost Configuration Section */}
+            {/* Connection Status Section */}
             <div className="mt-8 pt-6 border-t border-gray-200">
+              <MCPConnectionStatus
+                accessToken={accessToken}
+                formValues={formValues}
+                onToolsLoaded={setTools}
+              />
+            </div>
+
+            {/* Cost Configuration Section */}
+            <div className="mt-6">
               <MCPServerCostConfig
                 value={costConfig}
                 onChange={setCostConfig}
-                accessToken={accessToken}
+                tools={tools}
                 disabled={false}
-                formValues={formValues}
               />
             </div>
